Using default integers instead of fixed size integers. Prioritising cpu performence and not memory.

+++ DriveLogic.h
@@ -1,19 +1,18 @@
 #ifndef __DRIVE_LOGIC_H
 #define __DRIVE_LOGIC_H
 
-#include <stdint.h>
-#include <string>
 #include "HBridgeMotor.h"
+#include <string>
 
 
 class DriveLogic {
 public:
-  DriveLogic(double joystickMaxVal, int16_t joystickDeadzone, uint8_t maxMotorSpeed);
+  DriveLogic(double joystickMaxVal, unsigned int joystickDeadzone, unsigned int maxMotorSpeed);
   bool handleJoystickInput(
-    int16_t x, 
-    int16_t y, 
-    uint8_t& speedR, 
-    uint8_t& speedL, 
+    int x, 
+    int y, 
+    unsigned int& speedR, 
+    unsigned int& speedL, 
     HBridgeMotor::Direction& directionR, 
     HBridgeMotor::Direction& directionL
   );
@@ -23,11 +22,11 @@ private:
   //Constants//
   static const std::string kTag;
   const double kJoystickMaxVal;
-  const int16_t kJoystickDeadzone;
-  const uint8_t kMaxMotorSpeed;
+  const unsigned int kJoystickDeadzone;
+  const unsigned int kMaxMotorSpeed;
 
   //Methods//
-  bool joystickToSpeed(int16_t x, int16_t y, int16_t& speedR, int16_t& speedL);
-  HBridgeMotor::Direction speedToDirection(int16_t speed);
+  bool joystickToSpeed(int x, int y, int& speedR, int& speedL);
+  HBridgeMotor::Direction speedToDirection(int speed);
 };
 #endif
